rework to simplify
