+++ src/lib/logging.ts
@@ -1,38 +1,64 @@
+import Transport from "winston-transport";
+import { createDefaultTransportFormat } from "@zwave-js/core";
+import type { ZWaveLogInfo } from "@zwave-js/core";
 import { ClientsController, ZwavejsServer } from "./server";
+import { Driver } from "zwave-js";
 
 export class LoggingEventForwarder {
   /**
    * Only load this once the driver is ready.
    *
    * @param clients
-   * @param server
+   * @param driver
    */
+  private serverTransport?: EventEmitterLogTransport;
   public started: boolean = false;
 
-  constructor(
-    private clients: ClientsController,
-    private server: ZwavejsServer
-  ) {}
+  constructor(private clients: ClientsController, private driver: Driver) {
+    // Create log transport for server
+    this.serverTransport = new EventEmitterLogTransport(this.clients);
+  }
 
   start() {
-    // Forward logging events on to clients that are currently
-    // receiving logs
-    this.server.on("logging", (message: string) => {
-      this.clients.clients
-        .filter((cl) => cl.receiveLogs && cl.isConnected)
-        .forEach((client) =>
-          client.sendEvent({
-            source: "driver",
-            event: "logging",
-            message,
-          })
-        );
-    });
+    if (!this.serverTransport || this.serverTransport === undefined) {
+      throw new Error("Cannot start listening to logs");
+    }
+    const transports = this.driver.getLogConfig().transports || [];
+    transports.push(this.serverTransport);
+    this.driver.updateLogConfig({ transports });
     this.started = true;
   }
 
   stop() {
-    this.server.removeAllListeners("logging");
+    var transports = this.driver.getLogConfig().transports;
+    transports = transports.filter(
+      (transport) => transport !== this.serverTransport
+    );
+    this.driver.updateLogConfig({ transports });
     this.started = false;
+  }
+}
+
+class EventEmitterLogTransport extends Transport {
+  private messageSymbol: Symbol;
+
+  public constructor(private clients: ClientsController) {
+    super({ format: createDefaultTransportFormat(false, false) });
+    this.messageSymbol = Symbol.for("message");
+  }
+
+  public log(info: ZWaveLogInfo, next: () => void): any {
+    // Forward logs on to clients that are currently
+    // receiving logs
+    this.clients.clients
+      .filter((cl) => cl.receiveLogs && cl.isConnected)
+      .forEach((client) =>
+        client.sendEvent({
+          source: "driver",
+          event: "logging",
+          message: info[this.messageSymbol as any],
+        })
+      );
+    next();
   }
 }
