+++ src/lib/server.ts
@@ -1,10 +1,7 @@
-import Transport from "winston-transport";
 import ws from "ws";
 import type WebSocket from "ws";
 import type { Driver } from "zwave-js";
 import { libVersion } from "zwave-js";
-import { createDefaultTransportFormat } from "@zwave-js/core";
-import type { ZWaveLogInfo } from "@zwave-js/core";
 import { EventForwarder } from "./forward";
 import type * as OutgoingMessages from "./outgoing_message";
 import { IncomingMessage } from "./incoming_message";
@@ -219,11 +216,7 @@ export class ClientsController {
   private cleanupScheduled = false;
   private loggingEventForwarder?: LoggingEventForwarder;
 
-  constructor(
-    public driver: Driver,
-    private logger: Logger,
-    private server: ZwavejsServer
-  ) {}
+  constructor(public driver: Driver, private logger: Logger) {}
 
   addSocket(socket: WebSocket) {
     this.logger.debug("New client");
@@ -261,7 +254,7 @@ export class ClientsController {
     }
 
     if (this.loggingEventForwarder === undefined) {
-      this.loggingEventForwarder = new LoggingEventForwarder(this, this.server);
+      this.loggingEventForwarder = new LoggingEventForwarder(this, this.driver);
     }
   }
 
@@ -324,7 +317,6 @@ export interface ZwavejsServer {
   destroy(): void;
   on(event: "listening", listener: () => void): this;
   on(event: "error", listener: (error: Error) => void): this;
-  on(event: "logging", listener: (message: string) => void): this;
 }
 
 export class ZwavejsServer extends EventEmitter {
@@ -332,7 +324,6 @@ export class ZwavejsServer extends EventEmitter {
   private wsServer?: ws.Server;
   private sockets?: ClientsController;
   private logger: Logger;
-  private serverTransport?: EventEmitterLogTransport;
 
   constructor(private driver: Driver, private options: ZwavejsServerOptions) {
     super();
@@ -344,15 +335,9 @@ export class ZwavejsServer extends EventEmitter {
       throw new Error("Cannot start server when driver not ready");
     }
 
-    // Create log transport for server and update the drivers log configuration
-    this.serverTransport = new EventEmitterLogTransport(this);
-    const transports = this.driver.getLogConfig().transports || [];
-    transports.push(this.serverTransport);
-    this.driver.updateLogConfig({ transports });
-
     this.server = createServer();
     this.wsServer = new ws.Server({ server: this.server });
-    this.sockets = new ClientsController(this.driver, this.logger, this);
+    this.sockets = new ClientsController(this.driver, this.logger);
     this.wsServer.on("connection", (socket) => this.sockets!.addSocket(socket));
 
     this.logger.debug(`Starting server on port ${this.options.port}`);
@@ -380,19 +365,5 @@ export class ZwavejsServer extends EventEmitter {
     }
 
     this.logger.info(`Server closed`);
-  }
-}
-
-export class EventEmitterLogTransport extends Transport {
-  private messageSymbol: Symbol;
-
-  public constructor(private emitter: EventEmitter) {
-    super({ format: createDefaultTransportFormat(false, false) });
-    this.messageSymbol = Symbol.for("message");
-  }
-
-  public log(info: ZWaveLogInfo, next: () => void): any {
-    this.emitter.emit("logging", info[this.messageSymbol as any]);
-    next();
   }
 }
